+++ flink-runtime/src/test/java/org/apache/flink/runtime/state/filesystem/FsCheckpointStateOutputStreamTest.java
@@ -23,22 +23,29 @@ import org.apache.flink.core.fs.FSDataOutputStream;
 import org.apache.flink.core.fs.FileSystem;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.runtime.state.CheckpointStreamFactory;
+import org.apache.flink.runtime.state.CheckpointStreamFactory.CheckpointStateOutputStream;
 import org.apache.flink.runtime.state.StreamStateHandle;
 import org.apache.flink.runtime.state.filesystem.FsCheckpointStreamFactory.FsCheckpointStateOutputStream;
 import org.apache.flink.runtime.state.memory.ByteStreamStateHandle;
 
 import org.junit.Assert;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
 import org.mockito.ArgumentCaptor;
 
 import java.io.DataInputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URI;
 import java.util.Random;
+import java.util.function.Supplier;
 
 import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Matchers.any;
@@ -51,22 +58,23 @@ import static org.mockito.Mockito.when;
 
 public class FsCheckpointStateOutputStreamTest {
 
-	/** The temp dir, obtained in a platform neutral way */
-	private static final Path TEMP_DIR_PATH = new Path(new File(System.getProperty("java.io.tmpdir")).toURI());
+	@Rule
+	public final TemporaryFolder tempDir = new TemporaryFolder();
 
 
 	@Test(expected = IllegalArgumentException.class)
-	public void testWrongParameters() {
+	public void testWrongParameters() throws Exception {
 		// this should fail
 		new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(
-			TEMP_DIR_PATH, FileSystem.getLocalFileSystem(), 4000, 5000);
+			Path.fromLocalFile(tempDir.newFolder()), FileSystem.getLocalFileSystem(), 4000, 5000);
 	}
 
 
 	@Test
 	public void testEmptyState() throws Exception {
 		FsCheckpointStreamFactory.CheckpointStateOutputStream stream =
-				new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(TEMP_DIR_PATH, FileSystem.getLocalFileSystem(), 1024, 512);
+				new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(
+						Path.fromLocalFile(tempDir.newFolder()), FileSystem.getLocalFileSystem(), 1024, 512);
 
 		StreamStateHandle handle = stream.closeAndGetHandle();
 		assertTrue(handle == null);
@@ -95,7 +103,8 @@ public class FsCheckpointStateOutputStreamTest {
 	@Test
 	public void testGetPos() throws Exception {
 		FsCheckpointStreamFactory.CheckpointStateOutputStream stream =
-				new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(TEMP_DIR_PATH, FileSystem.getLocalFileSystem(), 31, 17);
+				new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(
+						Path.fromLocalFile(tempDir.newFolder()), FileSystem.getLocalFileSystem(), 31, 17);
 
 		for (int i = 0; i < 64; ++i) {
 			Assert.assertEquals(i, stream.getPos());
@@ -106,7 +115,8 @@ public class FsCheckpointStateOutputStreamTest {
 
 		// ----------------------------------------------------
 
-		stream = new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(TEMP_DIR_PATH, FileSystem.getLocalFileSystem(), 31, 17);
+		stream = new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(
+				Path.fromLocalFile(tempDir.newFolder()), FileSystem.getLocalFileSystem(), 31, 17);
 
 		byte[] data = "testme!".getBytes(ConfigConstants.DEFAULT_CHARSET);
 
@@ -133,7 +143,7 @@ public class FsCheckpointStateOutputStreamTest {
 		when(fs.create(pathCaptor.capture(), any(FileSystem.WriteMode.class))).thenReturn(outputStream);
 
 		CheckpointStreamFactory.CheckpointStateOutputStream stream = new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(
-			TEMP_DIR_PATH,
+			Path.fromLocalFile(tempDir.newFolder()),
 			fs,
 			4,
 			0);
@@ -162,7 +172,7 @@ public class FsCheckpointStateOutputStreamTest {
 		doThrow(new IOException("Test IOException.")).when(outputStream).close();
 
 		CheckpointStreamFactory.CheckpointStateOutputStream stream = new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(
-			TEMP_DIR_PATH,
+			Path.fromLocalFile(tempDir.newFolder()),
 			fs,
 			4,
 			0);
@@ -185,7 +195,7 @@ public class FsCheckpointStateOutputStreamTest {
 	private void runTest(int numBytes, int bufferSize, int threshold, boolean expectFile) throws Exception {
 		FsCheckpointStreamFactory.CheckpointStateOutputStream stream =
 			new FsCheckpointStreamFactory.FsCheckpointStateOutputStream(
-				TEMP_DIR_PATH, FileSystem.getLocalFileSystem(), bufferSize, threshold);
+					Path.fromLocalFile(tempDir.newFolder()), FileSystem.getLocalFileSystem(), bufferSize, threshold);
 
 		Random rnd = new Random();
 		byte[] original = new byte[numBytes];
@@ -233,7 +243,7 @@ public class FsCheckpointStateOutputStreamTest {
 	@Test
 	public void testWriteFailsFastWhenClosed() throws Exception {
 		FsCheckpointStateOutputStream stream = new FsCheckpointStateOutputStream(
-				TEMP_DIR_PATH, FileSystem.getLocalFileSystem(), 1024, 512);
+				Path.fromLocalFile(tempDir.newFolder()), FileSystem.getLocalFileSystem(), 1024, 512);
 
 		assertFalse(stream.isClosed());
 
@@ -252,6 +262,114 @@ public class FsCheckpointStateOutputStreamTest {
 			fail();
 		} catch (IOException e) {
 			// expected
+		}
+	}
+
+	@Test
+	public void testMixedBelowAndAboveThreshold() throws Exception {
+		final byte[] state1 = new byte[1274673];
+		final byte[] state2 = new byte[1];
+		final byte[] state3 = new byte[0];
+		final byte[] state4 = new byte[177];
+
+		final Random rnd = new Random();
+		rnd.nextBytes(state1);
+		rnd.nextBytes(state2);
+		rnd.nextBytes(state3);
+		rnd.nextBytes(state4);
+
+		final File directory = tempDir.newFolder();
+		final Path basePath = Path.fromLocalFile(directory);
+
+		final Supplier<CheckpointStateOutputStream> factory = () ->
+				new FsCheckpointStateOutputStream(basePath, FileSystem.getLocalFileSystem(), 1024, 15);
+
+		CheckpointStateOutputStream stream1 = factory.get();
+		CheckpointStateOutputStream stream2 = factory.get();
+		CheckpointStateOutputStream stream3 = factory.get();
+
+		stream1.write(state1);
+		stream2.write(state2);
+		stream3.write(state3);
+
+		FileStateHandle handle1 = (FileStateHandle) stream1.closeAndGetHandle();
+		ByteStreamStateHandle handle2 = (ByteStreamStateHandle) stream2.closeAndGetHandle();
+		ByteStreamStateHandle handle3 = (ByteStreamStateHandle) stream3.closeAndGetHandle();
+
+		// use with try-with-resources
+		StreamStateHandle handle4;
+		try (CheckpointStreamFactory.CheckpointStateOutputStream stream4 = factory.get()) {
+			stream4.write(state4);
+			handle4 = stream4.closeAndGetHandle();
+		}
+
+		// close before accessing handle
+		CheckpointStreamFactory.CheckpointStateOutputStream stream5 = factory.get();
+		stream5.write(state4);
+		stream5.close();
+		try {
+			stream5.closeAndGetHandle();
+			fail();
+		} catch (IOException e) {
+			// uh-huh
+		}
+
+		validateBytesInStream(handle1.openInputStream(), state1);
+		handle1.discardState();
+		assertFalse(isDirectoryEmpty(directory));
+		ensureLocalFileDeleted(handle1.getFilePath());
+
+		validateBytesInStream(handle2.openInputStream(), state2);
+		handle2.discardState();
+		assertFalse(isDirectoryEmpty(directory));
+
+		// nothing was written to the stream, so it will return nothing
+		assertNull(handle3);
+		assertFalse(isDirectoryEmpty(directory));
+
+		validateBytesInStream(handle4.openInputStream(), state4);
+		handle4.discardState();
+		assertTrue(isDirectoryEmpty(directory));
+	}
+
+	// ------------------------------------------------------------------------
+	//  Utilities
+	// ------------------------------------------------------------------------
+
+	private static void ensureLocalFileDeleted(Path path) {
+		URI uri = path.toUri();
+		if ("file".equals(uri.getScheme())) {
+			File file = new File(uri.getPath());
+			assertFalse("file not properly deleted", file.exists());
+		}
+		else {
+			throw new IllegalArgumentException("not a local path");
+		}
+	}
+
+	private static boolean isDirectoryEmpty(File directory) {
+		if (!directory.exists()) {
+			return true;
+		}
+		String[] nested = directory.list();
+		return nested == null || nested.length == 0;
+	}
+
+	private static void validateBytesInStream(InputStream is, byte[] data) throws IOException {
+		try {
+			byte[] holder = new byte[data.length];
+
+			int pos = 0;
+			int read;
+			while (pos < holder.length && (read = is.read(holder, pos, holder.length - pos)) != -1) {
+				pos += read;
+			}
+
+			assertEquals("not enough data", holder.length, pos);
+			assertEquals("too much data", -1, is.read());
+			assertArrayEquals("wrong data", data, holder);
+		} finally {
+			is.close();
 		}
 	}
 }
